% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_plan.R
\name{by_plan}
\alias{by_plan}
\title{Shorten District by Plan vector}
\usage{
by_plan(x, ndists)
}
\arguments{
\item{x}{summary statistic at the district level}

\item{ndists}{numeric. Number of districts. Estimated as the gcd of the unique
run length encodings if missing.}
}
\value{
x or plan level subset of x
}
\description{
If \code{x} is repeated for each district, it returns a plan level value. Otherwise
it returns \code{x}.
}
\examples{
by_plan(letters)
by_plan(rep(letters, each = 2))

}
