% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_bc}
\alias{comp_bc}
\title{Calculate Boyce Clark Ratio}
\usage{
comp_bc(plans, shp, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{A \code{redist_plans} object or plans_matrix where each row indicates a district assignment and each column is a plan.}

\item{shp}{A \code{redist_map} object, tibble, or data frame with an \code{sf} geometry column.}

\item{epsg}{Numeric EPSG code to use to project the shapefile, if needed. Default is 3857.}

\item{ncores}{Integer number of cores to use. Default is 1.}
}
\value{
A numeric vector. Can be shaped into a district-by-plan matrix.
}
\description{
Calculate Boyce Clark Ratio
}
\examples{
data(nh)
data(nh_m)
# For a single plan:
comp_bc(plans = nh$r_2020, shp = nh)

# Or many plans:
\donttest{
# slower, beware!
comp_bc(plans = nh_m[, 3:5], shp = nh)
}

}
\references{
Boyce, R., & Clark, W. 1964. The Concept of Shape in Geography.
Geographical Review, 54(4), 561-572.
}
\concept{compactness}
