% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_ch}
\alias{comp_ch}
\title{Calculate Convex Hull Compactness}
\usage{
comp_ch(shp, plans, epsg = 3857, ncores = 1)
}
\arguments{
\item{shp}{todo inherits}

\item{plans}{todo inherits}

\item{epsg}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Convex Hull Compactness
}
\examples{
# todo example
}
\concept{compactness}
