% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_holes}
\alias{comp_holes}
\title{Count District Holes}
\usage{
comp_holes(plans, shp, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{A \code{redist_plans} object or plans_matrix where each row indicates a district assignment and each column is a plan.}

\item{shp}{A \code{redist_map} object, tibble, or data frame with an \code{sf} geometry column.}

\item{epsg}{Numeric EPSG code to use to project the shapefile, if needed. Default is 3857.}

\item{ncores}{Integer number of cores to use. Default is 1.}
}
\value{
A numeric vector. Can be shaped into a district-by-plan matrix.
}
\description{
Counts interior rings after uniting each district geometry.
}
\examples{
data(nh)
data(nh_m)
comp_holes(plans = nh$r_2020, shp = nh)
comp_holes(plans = nh_m[, 1:2], shp = nh)
}
\concept{compactness}
