% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_lw}
\alias{comp_lw}
\title{Calculate Length Width Compactness}
\usage{
comp_lw(plans, shp, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
numeric vector
}
\description{
Calculate Length Width Compactness
}
\examples{
# todo example
}
\concept{compactness}
