% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_y_sym}
\alias{comp_y_sym}
\title{Calculate Y Symmetry Compactness}
\usage{
comp_y_sym(plans, shp, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
numeric vector
}
\description{
Y symmetry is the overlapping area of a shape and its projection over the
y-axis.
}
\examples{
#' data(nh)
data(nh_m)
# For a single plan:
comp_y_sym(plans = nh$r_2020, shp = nh)

# Or many plans:
\donttest{
# slower, beware!
comp_y_sym(plans = nh_m[, 3:5], shp = nh)
}

}
\references{
Aaron Kaufman, Gary King, and Mayya Komisarchik. 2021.
How to Measure Legislative District COmpactness If You Only Know it When You See It.
American Journal of Political Science. 65, 3. Pp. 533-550.
}
\concept{compactness}
