% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_info}
\alias{dist_info}
\title{Calculate Variation of Information Distances}
\usage{
dist_info(plans, total_pop, ncores = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{total_pop}{A numeric vector with the population for every observation.}

\item{ncores}{todo inherits}
}
\value{
matrix of plan distances
}
\description{
Calculate Variation of Information Distances
}
\examples{
# todo example
}
