% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_man}
\alias{dist_man}
\title{Calculate Manhattan Distances}
\usage{
dist_man(plans, ncores = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
matrix of plan distances
}
\description{
Calculate Manhattan Distances
}
\examples{
data(nh)
data(nh_m)
# For a single plan (distance is trivial, 0):
dist_man(plans = nh$r_2020)

# Or many plans:
dist_man(plans = nh_m[, 3:5])

}
\concept{distances}
