% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incumbent.R
\name{inc_pairs}
\alias{inc_pairs}
\title{Count Incumbent Pairings}
\usage{
inc_pairs(plans, shp, inc)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{inc}{unqouted name of logical column in shp indicating where incumbents live}
}
\value{

}
\description{
Count Incumbent Pairings
}
\examples{
# todo example
}
\concept{incumbent}
