% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fn.R
\name{list_fn}
\alias{list_fn}
\title{Return Functions Matching a Prefix}
\usage{
list_fn(prefix)
}
\arguments{
\item{prefix}{character prefix of functions to return}
}
\value{
character vector of functions
}
\description{
This package uses prefixes for each function that correspond to the type of measure.
This function returns the functions
}
\examples{
list_fn('part_')
}
