% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nh_m}
\alias{nh_m}
\title{Redistricting Plans for New Hampshire as \code{matrix}}
\format{
A matrix with 52 columns and 326 rows where each column is a plan
}
\usage{
data("nh_m")
}
\description{
This data set contains two reference plans (\code{d_2020} and \code{r_2020}) and
50 simulated plans for New Hampshire, based on 2020 demographics, simulated at
a population tolerance of 0.05\%.
}
\examples{
data(nh_m)

}
\concept{data}
