% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nh_plans}
\alias{nh_plans}
\title{Redistricting Plans for New Hampshire as \code{redist_plans}}
\format{
A redist_plans with 104 rows and 3 columns
\itemize{
\item draw: factor identifying the reference plans (\code{d_2020} and \code{r_2020}) and 50 simulted plans
\item district: district number (\code{1} or \code{2})
\item total_pop: total population in the district
}
}
\usage{
data("nh_plans")
}
\description{
This data set contains two reference plans (\code{d_2020} and \code{r_2020}) and
50 simulated plans for New Hampshire, based on 2020 demographics, simulated at
a population tolerance of 0.05\%.
}
\examples{
data(nh_plans)

}
\concept{data}
