% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_egap_ep}
\alias{part_egap_ep}
\title{Calculate Efficiency Gap (Equal Population Assumption)}
\usage{
part_egap_ep(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Efficiency Gap (Equal Population Assumption)
}
\examples{
# todo example
}
\concept{partisan}
