% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_mean_median}
\alias{part_mean_median}
\title{Calculate Mean Median Score}
\usage{
part_mean_median(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Mean Median Score
}
\examples{
# todo example
}
\concept{partisan}
