% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_resp}
\alias{part_resp}
\title{Calculate Responsiveness}
\usage{
part_resp(plans, dvote, rvote, v, bandwidth = FALSE)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}

\item{v}{vote share to calculate bias at. Numeric. Default is 0.5.}

\item{bandwidth}{Defaults to 0.01. A value between 0 and 1 for the step size to estimate the slope.}
}
\value{
numeric vector
}
\description{
Calculate Responsiveness
}
\examples{
# todo example
}
\concept{partisan}
