% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_rmd}
\alias{part_rmd}
\title{Calculate Ranked Margin Deviation}
\usage{
part_rmd(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Ranked Margin Deviation
}
\examples{
# todo example
}
\concept{partisan}
