% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness_helpers.R
\name{prep_perims}
\alias{prep_perims}
\title{Prep Polsby Popper Perimeter Tibble}
\usage{
prep_perims(shp, epsg = 3857, perim_path, ncores = 1)
}
\arguments{
\item{shp}{redist_map object or tibble with sf geometry column}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
tibble of perimeters and lengths
}
\description{
Replaces \code{redist.prep.polsbypopper}
}
\examples{
# todo example
}
