% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\name{seg_dissim}
\alias{seg_dissim}
\title{Compute Dissimilarity Index}
\usage{
seg_dissim(plans, shp, group_pop, total_pop)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{group_pop}{unqouted name of column in shp with group population}

\item{total_pop}{unqouted name of column in shp with total population}
}
\value{
numeric vector
}
\description{
Compute Dissimilarity Index
}
\examples{
# todo examples
}
\concept{segregation}
