% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{splits_sub_admin}
\alias{splits_sub_admin}
\title{Compute Number of Sub-Administrative Units Split More than Once}
\usage{
splits_sub_admin(plans, shp, sub_admin)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{sub_admin}{todo inherits}
}
\value{
numeric vector
}
\description{
Compute Number of Sub-Administrative Units Split More than Once
}
\examples{
# todo examples
}
\concept{splits}
