% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{splits_sub_count}
\alias{splits_sub_count}
\title{Count the Number of Splits in Each Sub-Administrative Unit}
\usage{
splits_sub_count(plans, shp, sub_admin)
}
\arguments{
\item{plans}{A \code{redist_plans} object or plans_matrix where each row indicates a district assignment and each column is a plan.}

\item{shp}{A \code{redist_map} object, tibble, or data frame containing other columns.}

\item{sub_admin}{Unqouted name of column in shp with numeric identifiers for subsidiary administrative units.}
}
\value{
numeric matrix
}
\description{
Tallies the number of unique sub-administrative unit-districts. An unsplit administrative
unit will return an entry of 1, while each additional sub-administrative unit-district
adds 1.
}
\examples{
data(nh)
data(nh_m)
# For a single plan:
splits_sub_count(plans = nh$r_2020, shp = nh, sub_admin = county)

# Or many plans:
splits_sub_count(plans = nh_m[, 3:5], shp = nh, sub_admin = county)

}
\concept{splits}
