% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_bc}
\alias{comp_bc}
\title{Calculate Boyce Clark Ratio}
\usage{
comp_bc(plans, shp, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{epsg}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Boyce Clark Ratio
}
\examples{
# todo example
}
\concept{compactness}
