% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_edges_rem}
\alias{comp_edges_rem}
\title{Calulate Edges Removed Compactness}
\usage{
comp_edges_rem(plans, shp, adj)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{adj}{zero-indexed adjacency list. Not required if \code{redist_map} supplied to \code{shp}.}
}
\value{
numeric vector
}
\description{
Calulate Edges Removed Compactness
}
\examples{
# todo example
}
\concept{compactness}
