% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_edges_rem}
\alias{comp_edges_rem}
\title{Calulate Edges Removed Compactness}
\usage{
comp_edges_rem(plans, shp, adj)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{adj}{todo inherits}
}
\value{
numeric vector
}
\description{
Calulate Edges Removed Compactness
}
\examples{
# todo example
}
\concept{compactness}
