% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_fh}
\alias{comp_fh}
\title{Calulate Fryer Holden Compactness}
\usage{
comp_fh(shp, plans, total_pop, epsg = 3857, ncores = 1)
}
\arguments{
\item{shp}{todo inherits}

\item{plans}{todo inherits}

\item{total_pop}{A numeric vector with the population for every observation.}

\item{epsg}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
numeric vector
}
\description{
Calulate Fryer Holden Compactness
}
\examples{
# todo example
}
\concept{compactness}
