% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_fh}
\alias{comp_fh}
\title{Calulate Fryer Holden Compactness}
\usage{
comp_fh(plans, shp, total_pop, epsg = 3857, ncores = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{total_pop}{A numeric vector with the population for every observation.}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{ncores}{TRUE}
}
\value{
numeric vector
}
\description{
Calulate Fryer Holden Compactness
}
\examples{
# todo example
}
\concept{compactness}
