% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_frac_kept}
\alias{comp_frac_kept}
\title{Calulate Fraction Kept Compactness}
\usage{
comp_frac_kept(plans, shp, adj)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{adj}{todo inherits}
}
\value{
numeric vector
}
\description{
Calulate Fraction Kept Compactness
}
\examples{
# todo example
}
\concept{compactness}
