% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_frac_kept}
\alias{comp_frac_kept}
\title{Calulate Fraction Kept Compactness}
\usage{
comp_frac_kept(plans, shp, adj)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{adj}{zero-indexed adjacency list. Not required if \code{redist_map} supplied to \code{shp}.}
}
\value{
numeric vector
}
\description{
Calulate Fraction Kept Compactness
}
\examples{
data(nh)
data(nh_m)
# For a single plan:
comp_frac_kept(plans = nh$r_2020, shp = nh, nh$adj)

# Or many plans:
comp_frac_kept(plans = nh_m[, 3:5], shp = nh, nh$adj)

}
\concept{compactness}
