% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwysi.R
\name{comp_kiwysi}
\alias{comp_kiwysi}
\title{Calculate KIWSI Compactness}
\usage{
comp_kiwysi(plans, shp, epsg = 3857, ncores = 1, return_all = FALSE, ...)
}
\arguments{
\item{plans}{A \code{redist_plans} object or plans_matrix where each row indicates a district assignment and each column is a plan.}

\item{shp}{A \code{redist_map} object, tibble, or data frame with an \code{sf} geometry column.}

\item{epsg}{Numeric EPSG code to use to project the shapefile, if needed. Default is 3857.}

\item{ncores}{Integer number of cores to use. Default is 1.}

\item{return_all}{If \code{TRUE}, return a tibble containing the score and all
features used by the model.}

\item{...}{Precomputed feature vectors. These are tidy-evaluated in \code{shp}.
Use the corresponding function name, such as \code{comp_polsby = polsby}.}
}
\value{
A numeric vector. Can be shaped into a district-by-plan matrix.
}
\description{
Predicts the Kaufman, King, and Komisarchik "know it when you see it"
compactness score from \code{redistmetrics} compactness measures. Scores are
bounded between 0 and 100, where larger values indicate less compact
districts.
}
\examples{
data(nh)
comp_kiwysi(plans = nh$r_2020, shp = nh)
}
\references{
Aaron Kaufman, Gary King, and Mayya Komisarchik. 2021.
How to Measure Legislative District Compactness If You Only Know it When You See It.
American Journal of Political Science. 65, 3. Pp. 533-550.
}
\concept{compactness}
