% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_log_st}
\alias{comp_log_st}
\title{Calulate Log Spanning Tree Compactness}
\usage{
comp_log_st(plans, shp, counties, adj)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{counties}{column name in shp containing counties}

\item{adj}{todo inherits}
}
\value{
numeric vector
}
\description{
Calulate Log Spanning Tree Compactness
}
\examples{
# todo example
}
\concept{compactness}
