% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_log_st}
\alias{comp_log_st}
\title{Calulate Log Spanning Tree Compactness}
\usage{
comp_log_st(plans, shp, counties, adj)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{counties}{column name in shp containing counties}

\item{adj}{zero-indexed adjacency list. Not required if \code{redist_map} supplied to \code{shp}.}
}
\value{
numeric vector
}
\description{
Calulate Log Spanning Tree Compactness
}
\examples{
# todo example
}
\concept{compactness}
