% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_lw}
\alias{comp_lw}
\title{Calculate Length Width Compactness}
\usage{
comp_lw(shp, plans, epsg = 3857, ncores = 1)
}
\arguments{
\item{shp}{todo inherits}

\item{plans}{todo inherits}

\item{epsg}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Length Width Compactness
}
\examples{
# todo example
}
\concept{compactness}
