% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_polsby}
\alias{comp_polsby}
\title{Calculate Polsby Popper Compactness}
\usage{
comp_polsby(
  plans,
  shp,
  use_Rcpp,
  perim_path,
  perim_df,
  epsg = 3857,
  ncores = 1
)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{use_Rcpp}{Logical. Use Rcpp?}

\item{perim_path}{path to perimeter tibble saved by \code{prep_perims()}}

\item{perim_df}{tibble of perimeters from \code{prep_perims()}}

\item{epsg}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Polsby Popper Compactness
}
\examples{
# todo example
}
\concept{compactness}
