% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_reock}
\alias{comp_reock}
\title{Calculate Reock Compactness}
\usage{
comp_reock(shp, plans, epsg = 3857, ncores = 1)
}
\arguments{
\item{shp}{todo inherits}

\item{plans}{todo inherits}

\item{epsg}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Reock Compactness
}
\examples{
# todo example
}
\concept{compactness}
