% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness.R
\name{comp_schwartz}
\alias{comp_schwartz}
\title{Calculate Schwartzberg Compactness}
\usage{
comp_schwartz(
  plans,
  shp,
  use_Rcpp,
  perim_path,
  perim_df,
  epsg = 3857,
  ncores = 1
)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{use_Rcpp}{Logical. Use Rcpp?}

\item{perim_path}{path to perimeter tibble saved by \code{prep_perims()}}

\item{perim_df}{tibble of perimeters from \code{prep_perims()}}

\item{epsg}{numeric EPSG code to planarize to. Default is 3857.}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
numeric vector
}
\description{
Calculate Schwartzberg Compactness
}
\examples{
# todo example
}
\concept{compactness}
