% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness.R
\name{compet_talisman}
\alias{compet_talisman}
\title{Compute Talismanic Redistricting Competitiveness Metric}
\usage{
compet_talisman(plans, rvote, dvote, alpha = 1, beta = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{rvote}{todo inherits}

\item{dvote}{todo inherits}

\item{alpha}{Numeric scaling value}

\item{beta}{Numeric scaling value}
}
\value{
numeric vector
}
\description{
Compute Talismanic Redistricting Competitiveness Metric
}
\examples{
# todo examples
}
