% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitiveness.R
\name{compet_talisman}
\alias{compet_talisman}
\title{Compute Talismanic Redistricting Competitiveness Metric}
\usage{
compet_talisman(plans, shp, rvote, dvote, alpha = 1, beta = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{rvote}{unqouted name of column in shp with group population}

\item{dvote}{unqouted name of column in shp with total population}

\item{alpha}{Numeric scaling value}

\item{beta}{Numeric scaling value}
}
\value{
numeric vector
}
\description{
Compute Talismanic Redistricting Competitiveness Metric
}
\examples{
data(nh)
data(nh_m)
# For a single plan:
compet_talisman(plans = nh$r_2020, shp = nh, rvote = nrv, dvote = ndv)

# Or many plans:
compet_talisman(plans = nh_m[, 3:5], shp = nh, rvote = nrv, dvote = ndv)

}
\references{
Wendy K. Tam Cho and Yan Y. Liu
Toward a Talismanic Redistricting Tool.
Election Law Journal. 15, 4. Pp. 351-366.
}
\concept{competitiveness}
