% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_euc}
\alias{dist_euc}
\title{Calculate Euclidean Distances}
\usage{
dist_euc(plans, ncores = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
matrix of plan distances
}
\description{
Calculate Euclidean Distances
}
\examples{
# todo example
}
