% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_ham}
\alias{dist_ham}
\title{Calculate Hamming Distances}
\usage{
dist_ham(plans, ncores = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
matrix of plan distances
}
\description{
Calculate Hamming Distances
}
\examples{
# todo example
}
