% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_ham}
\alias{dist_ham}
\title{Calculate Hamming Distances}
\usage{
dist_ham(plans, ncores = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{ncores}{numeric. Number of cores to use. Default is 1.}
}
\value{
matrix of plan distances
}
\description{
Calculate Hamming Distances
}
\examples{
# todo example
}
\concept{distances}
