% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_man}
\alias{dist_man}
\title{Calculate Manhattan Distances}
\usage{
dist_man(plans, ncores = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{ncores}{todo inherits}
}
\value{
matrix of plan distances
}
\description{
Calculate Manhattan Distances
}
\examples{
# todo example
}
\concept{distances}
