% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_bias}
\alias{part_bias}
\title{Calculate Partisan Bias}
\usage{
part_bias(plans, dvote, rvote, v = 0.5)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}

\item{v}{vote share to calculate bias at. Numeric. Default is 0.5.}
}
\value{
numeric vector
}
\description{
Calculate Partisan Bias
}
\examples{
# todo example
}
\concept{partisan}
