% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_bias}
\alias{part_bias}
\title{Calculate Partisan Bias}
\usage{
part_bias(plans, shp, dvote, rvote, v = 0.5)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{dvote}{unqouted name of column in shp with total population}

\item{rvote}{unqouted name of column in shp with group population}

\item{v}{vote share to calculate bias at. Numeric. Default is 0.5.}
}
\value{
numeric vector
}
\description{
Calculate Partisan Bias
}
\examples{
# todo example
}
\concept{partisan}
