% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_decl}
\alias{part_decl}
\title{Calculate Declination}
\usage{
part_decl(plans, dvote, rvote, normalize = FALSE)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}

\item{normalize}{Default is FALSE. Translate score to an angle?}
}
\value{
numeric vector
}
\description{
Calculate Declination
}
\examples{
# todo example
}
\concept{partisan}
