% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_decl}
\alias{part_decl}
\title{Calculate Declination}
\usage{
part_decl(plans, shp, dvote, rvote, normalize = FALSE)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{dvote}{unqouted name of column in shp with total population}

\item{rvote}{unqouted name of column in shp with group population}

\item{normalize}{Default is FALSE. Translate score to an angle?}
}
\value{
numeric vector
}
\description{
Calculate Declination
}
\examples{
# todo example
}
\concept{partisan}
