% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_dseats}
\alias{part_dseats}
\title{Calculate Democratic Seats}
\usage{
part_dseats(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Democratic Seats
}
\examples{
# todo example
}
\concept{partisan}
