% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_dvs}
\alias{part_dvs}
\title{Calculate Democratic Vote Share}
\usage{
part_dvs(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Democratic Vote Share
}
\examples{
# todo example
}
\concept{partisan}
