% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_egap}
\alias{part_egap}
\title{Calculate Efficiency Gap}
\usage{
part_egap(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Efficiency Gap
}
\examples{
# todo example
}
\concept{partisan}
