% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_egap}
\alias{part_egap}
\title{Calculate Efficiency Gap}
\usage{
part_egap(plans, shp, dvote, rvote)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{dvote}{unqouted name of column in shp with total population}

\item{rvote}{unqouted name of column in shp with group population}
}
\value{
numeric vector
}
\description{
Calculate Efficiency Gap
}
\examples{
# todo example
}
\concept{partisan}
