% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_lopsided_wins}
\alias{part_lopsided_wins}
\title{Calculate Lopsided Wins}
\usage{
part_lopsided_wins(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Lopsided Wins
}
\examples{
# todo example
}
\concept{partisan}
