% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_rmd}
\alias{part_rmd}
\title{Calculate Ranked Margin Deviation}
\usage{
part_rmd(plans, dvote, rvote)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{dvote}{unqouted name of column in shp with total population}

\item{rvote}{unqouted name of column in shp with group population}
}
\value{
numeric vector
}
\description{
Calculate Ranked Margin Deviation
}
\examples{
# todo example
}
\concept{partisan}
