% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_sscd}
\alias{part_sscd}
\title{Calculate Smoothed Seat Count Deviation}
\usage{
part_sscd(plans, dvote, rvote)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}
}
\value{
numeric vector
}
\description{
Calculate Smoothed Seat Count Deviation
}
\examples{
# todo example
}
\concept{partisan}
