% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_tau_gap}
\alias{part_tau_gap}
\title{Calculate Tau Gap}
\usage{
part_tau_gap(plans, dvote, rvote, tau = 1)
}
\arguments{
\item{plans}{todo inherits}

\item{dvote}{todo inherits}

\item{rvote}{todo inherits}

\item{tau}{A non-negative numeric for calculating Tau Gap. Defaults to 1.}
}
\value{
numeric vector
}
\description{
Calculate Tau Gap
}
\examples{
# todo example
}
\concept{partisan}
