% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partisan.R
\name{part_tau_gap}
\alias{part_tau_gap}
\title{Calculate Tau Gap}
\usage{
part_tau_gap(plans, dvote, rvote, tau = 1)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{dvote}{unqouted name of column in shp with total population}

\item{rvote}{unqouted name of column in shp with group population}

\item{tau}{A non-negative numeric for calculating Tau Gap. Defaults to 1.}
}
\value{
numeric vector
}
\description{
Calculate Tau Gap
}
\examples{
# todo example
}
\concept{partisan}
