% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactness_helpers.R
\name{prep_perims}
\alias{prep_perims}
\title{Prep Polsby Popper Perimeter Tibble}
\usage{
prep_perims(shp, epsg = 3857, perim_path, ncores = 1)
}
\arguments{
\item{perim_path}{A path to save an rds}
}
\value{
tibble of perimeters and lengths
}
\description{
Replaces \code{redist.prep.polsbypopper}
}
\examples{
data(nh)
prep_perims(nh)

}
\concept{compactness}
