% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregation.R
\name{seg_dissim}
\alias{seg_dissim}
\title{Compute Dissimilarity Index}
\usage{
seg_dissim(plans, shp, group_pop, total_pop)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{group_pop}{todo inherits}

\item{total_pop}{todo inherits}
}
\value{
numeric vector
}
\description{
Compute Dissimilarity Index
}
\examples{
# todo examples
}
\concept{segregation}
