% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{splits_admin}
\alias{splits_admin}
\title{Compute Number of Administrative Units Split}
\usage{
splits_admin(plans, shp, admin)
}
\arguments{
\item{plans}{todo inherits}

\item{shp}{todo inherits}

\item{admin}{todo inherits}
}
\value{
numeric vector
}
\description{
Compute Number of Administrative Units Split
}
\examples{
# todo examples
}
\concept{splits}
