% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{splits_sub_admin}
\alias{splits_sub_admin}
\title{Compute Number of Sub-Administrative Units Split}
\usage{
splits_sub_admin(plans, shp, sub_admin)
}
\arguments{
\item{plans}{redist_plans object or plans_matrix where each row indicates a district assignment and each column is a plan}

\item{shp}{redist_map object or tibble with sf geometry column}

\item{sub_admin}{unqouted name of column in shp with numeric identifiers for subsidiary administrative units}
}
\value{
numeric vector
}
\description{
Compute Number of Sub-Administrative Units Split
}
\examples{
# todo examples
}
\concept{splits}
